Move sv1_server to mod.rs
